requirements.txt
Flask
flask_cors
yahooquery
pandas
gunicorn
finnhub-python